Restore message update
Updated the messages if/else statements so that if a message is removed, you won't see the flag option when going to restore it.

One more change will be necessary at some point to change some of these if/elses to a switch. I used if/else in the meantime as I'm more familiar with that whilst I was working out the functionality